Fixed dashboard component usage.
